api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "pillow>=12.1.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=40.4.0",
]
